@@ -8,9 +8,10 @@
         "php": ">=7.1.3",
         "doctrine/dbal": "^2.6",
         "guzzlehttp/guzzle": "^6.3",
-        "laravel/framework": "5.7.*",
+        "laravel/framework": "5.8.*",
         "laravel/tinker": "~1.0",
         "league/flysystem-aws-s3-v3": "^1.0",
+        "matt-bartlett/php-spotify-api": "^1.0",
         "spatie/laravel-backup": "^5.2",
         "spatie/laravel-fractal": "^5.2"
     },
@@ -18,7 +19,7 @@
         "filp/whoops": "~2.0",
         "fzaninotto/faker": "~1.4",
         "mockery/mockery": "0.9.*",
-        "phpunit/phpunit": "~6.0",
+        "phpunit/phpunit": "^7.0",
         "squizlabs/php_codesniffer": "3.*"
     },
     "autoload": {
