@@ -12,32 +12,31 @@ class PlaylistTransformer extends Transformer implements TransformerInterface
      * and playlists.
      *
      * @param stdClass $object
+     *
      * @return array
      */
-    public function transform(stdClass $object)
+    public function transform(stdClass $object) : array
     {
-        $playlist = array();
+        $playlist = [];
         $playlist['name'] = $this->get($object, 'name');
         $playlist['owner_id'] = $this->get($object->owner, 'id');
         $playlist['playlist_url'] = $this->get($object->external_urls, 'spotify');
         $playlist['owner_profile_url'] = $this->get($object->owner->external_urls, 'spotify');
         $playlist['playlist_thumbnail_url'] = $this->get($object->images, 'url');
 
-        $tracks = array();
-        foreach ($object->tracks->items as $track) {
-            $trackArray = array();
-            $trackArray['title'] = $this->get($track->track, 'name');
-            $trackArray['album'] = $this->get($track->track->album, 'name');
-            $trackArray['artist'] = $this->get($track->track->album->artists, 'name');
-            $trackArray['duration'] = $this->get($track->track, 'duration_ms');
-            $trackArray['spotify_url'] = $this->get($track->track->external_urls, 'spotify');
-            $trackArray['spotify_track_id'] = $this->get($track->track, 'id');
-            $trackArray['spotify_preview_url'] = $this->get($track->track, 'preview_url');
-            $trackArray['spotify_thumbnail_url'] = $this->get($track->track->album->images, 'url');
-            $tracks[] = $trackArray;
-        }
-
-        $playlist['tracks'] = $tracks;
+        $playlist['tracks'] = collect($object->tracks->items)
+            ->map(function ($track) {
+                return [
+                    'title' => $this->get($track->track, 'name'),
+                    'album' => $this->get($track->track->album, 'name'),
+                    'artist' => $this->get($track->track->album->artists, 'name'),
+                    'duration' => $this->get($track->track, 'duration_ms'),
+                    'spotify_url' => $this->get($track->track->external_urls, 'spotify'),
+                    'spotify_track_id' => $this->get($track->track, 'id'),
+                    'spotify_preview_url' => $this->get($track->track, 'preview_url'),
+                    'spotify_thumbnail_url' => $this->get($track->track->album->images, 'url'),
+                ];
+            });
 
         return $playlist;
     }
