@@ -10,6 +10,8 @@ use App\Http\Requests\PlaylistRequest;
 use App\T3\Services\CreatePlaylistService;
 use App\Http\Requests\SpotifyPlaylistRequest;
 
+use Spotify\Resources\Playlist as SpotfiyPlaylist;
+
 class PlaylistController extends Controller
 {
     /**
@@ -65,13 +67,14 @@ class PlaylistController extends Controller
      *
      * @param App\Http\Requests\SpotifyPlaylistRequest $request
      * @param App\T3\Services\CreatePlaylistService $service
+     *
      * @return Illuminate\Http\RedirectResponse
      */
     public function store(SpotifyPlaylistRequest $request, CreatePlaylistService $service)
     {
         try {
             // Create Playlist based on input data
-            $service->make($request);
+            $service->handle($request);
 
             return redirect()->route('admin.playlists.index');
         } catch (Exception $e) {
