@@ -2,9 +2,14 @@
 
 namespace App\Providers;
 
+use Spotify\Http\Request;
+use Spotify\Auth\Credentials;
+use Illuminate\Session\Store;
+use Spotify\Contracts\Store\Session;
 use Illuminate\Support\ServiceProvider;
-use App\T3\Utils\Format\TrackDurationFormatter;
-use App\T3\Utils\Format\PlaylistDurationFormatter;
+use Spotify\Auth\Flows\ClientCredentials;
+use Spotify\Contracts\Auth\Authenticator;
+use Spotify\Sessions\LaravelSessionHandler;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -25,14 +30,28 @@ class AppServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        // Bind TrackDurationFormatter to Container
-        $this->app->bind('TrackDurationFormatter', function ($app) {
-            return new TrackDurationFormatter;
+        // Bind Spotify credentials.
+        $this->app->singleton(Credentials::class, function ($app) {
+            return new Credentials(
+                getenv('SPOTIFY_CLIENT_ID'),
+                getenv('SPOTIFY_CLIENT_SECRET'),
+                getenv('SPOTIFY_REDIRECT_URL')
+            );
         });
 
-        // Bind PlaylistDurationFormatter to Container
-        $this->app->bind('PlaylistDurationFormatter', function ($app) {
-            return new PlaylistDurationFormatter;
+        // Bind Client Credentials to the Spotify authenticator.
+        $this->app->bind(Authenticator::class, function ($app) {
+            return new ClientCredentials(
+                $app->make(Request::class),
+                $app->make(Credentials::class)
+            );
+        });
+
+        // Bind the Laravel Session handler.
+        $this->app->bind(Session::class, function ($app) {
+            return new LaravelSessionHandler(
+                $app->make(Store::class)
+            );
         });
     }
 }
