@@ -1,6 +1,6 @@
 <?php
 
-namespace App\T3\Utils\Format;
+namespace App\T3\Formatters;
 
 class PlaylistDurationFormatter
 {
@@ -8,11 +8,12 @@ class PlaylistDurationFormatter
      * Calculate the total duration of all tracks within a playlist
      *
      * @param Illuminate\Database\Eloquent\Collection $tracks
-     * @return integer
+     *
+     * @return int
      */
     public function format($tracks) : int
     {
-        $duration = $tracks->reduce(function ($carry, $track) {
+        $duration = $tracks->reduce(function (int $carry, $track) {
             return $carry + $track->duration;
         }, 0);
 
