@@ -5,6 +5,7 @@ namespace App\T3\Transformers;
 use App\Models\Playlist;
 use League\Fractal\TransformerAbstract;
 use App\T3\Transformers\TrackTransformer;
+use App\T3\Formatters\PlaylistDurationFormatter;
 use League\Fractal\Pagination\IlluminatePaginatorAdapter;
 
 class PlaylistTransformer extends TransformerAbstract
@@ -22,15 +23,16 @@ class PlaylistTransformer extends TransformerAbstract
      * Transform model output
      *
      * @param App\Models\Playlist $playlist
+     *
      * @return array
      */
-    public function transform(Playlist $playlist)
+    public function transform(Playlist $playlist) : array
     {
         return [
-            'id' => (integer) $playlist->id,
+            'id' => (int) $playlist->id,
             'name' => (string) $playlist->name,
             'owner' => (string) $playlist->owner_id,
-            'duration' => (integer) $playlist->duration,
+            'duration' => $this->formatPlaylistDuration($playlist->tracks),
             'owner_name' => (string) $playlist->owner_name,
             'playlist_url' => (string) $playlist->playlist_url,
             'owner_profile_url' => (string) $playlist->owner_profile_url,
@@ -42,6 +44,7 @@ class PlaylistTransformer extends TransformerAbstract
      * Transform model output
      *
      * @param App\Models\Playlist $playlist
+     *
      * @return League\Fractal\Resource\Collection
      */
     public function includeTracks(Playlist $playlist)
@@ -52,5 +55,17 @@ class PlaylistTransformer extends TransformerAbstract
             ->setPaginator(new IlluminatePaginatorAdapter($tracks));
 
         return $resource;
+    }
+
+    /**
+     * Calculate the total duration of all tracks within a playlist
+     *
+     * @param Illuminate\Database\Eloquent\Collection $tracks
+     *
+     * @return int
+     */
+    protected function formatPlaylistDuration($tracks) : int
+    {
+        return (new PlaylistDurationFormatter)->format($tracks);
     }
 }
