@@ -1,16 +1,16 @@
 <?php
 
-namespace Tests\Unit\Utils\Format;
+namespace Tests\Unit\Formatters;
 
 use Tests\TestCase;
 use App\Models\Track;
-use App\T3\Utils\Format\TrackDurationFormatter;
+use App\T3\Formatters\TrackDurationFormatter;
 
 class TrackDurationFormatterTest extends TestCase
 {
     public $formatter;
 
-    public function setUp()
+    public function setUp() : void
     {
         parent::setUp();
         $this->formatter = new TrackDurationFormatter;
@@ -19,7 +19,7 @@ class TrackDurationFormatterTest extends TestCase
     /**
      * @return void
      */
-    public function test_duration_formats()
+    public function test_duration_formats() : void
     {
         $this->assertEquals('0:53', $this->formatter->format(53000));
         $this->assertEquals('0:53', $this->formatter->format(53500));
@@ -30,7 +30,7 @@ class TrackDurationFormatterTest extends TestCase
     /**
      * @return void
      */
-    public function test_equally_divisible_duration_formats_an_exact_minute()
+    public function test_equally_divisible_duration_formats_an_exact_minute() : void
     {
         $this->assertEquals('3:00', $this->formatter->format(180000));
     }
@@ -38,7 +38,7 @@ class TrackDurationFormatterTest extends TestCase
     /**
      * @return void
      */
-    public function test_duration_formats_to_minutes()
+    public function test_duration_formats_to_minutes() : void
     {
         $this->assertEquals(3, $this->formatter->formatToMinutes(180000));
         $this->assertEquals(3, $this->formatter->formatToMinutes(230000));
